Clarify iOS installation in README to avoid errors
The iOS install instructions were not really clear and one could easily read it as you have to do the `iOS` section of the `Manual Installation`, as well as the `Using CocoaPods` section. Understanding it as: "Well, I'm using iOS, so gotta do that. Oh, and also, I'm using CocoaPods so I probably have to do that too to make it work". This way it's a clear if-else between using CocoaPods or not.